@@ -12,7 +12,7 @@ import Foundation
 /// Implementations may have different constraints and are free to convert entry contexts to their
 /// own subtypes. This means callers cannot assume the getCurrentContext()
 /// is the same instance as the one withContext() placed into scope.
-public protocol BaggageManager: AnyObject {
+public protocol BaggageManager: AnyObject, Sendable {
   /// Returns a new ContextBuilder.
   func baggageBuilder() -> BaggageBuilder
 }
