@@ -5,14 +5,15 @@
 
 import Foundation
 
-open class AttributeArray: Hashable, Codable {
+open class AttributeArray: Hashable, Codable, @unchecked Sendable {
   public private(set) var values: [AttributeValue]
-  public static var empty = AttributeArray()
+  @available(*, deprecated, message: "Create new instances instead of using shared instance")
+  public static let empty = AttributeArray()
   public var description: String {
     values.description
   }
 
-  private init() {
+  public init() {
     values = [AttributeValue]()
   }
 
