@@ -6,8 +6,11 @@
 import Foundation
 
 /// No-op implementation of the Tracer
-public class DefaultTracer: Tracer {
-  public static var instance = DefaultTracer()
+public final class DefaultTracer: Tracer, @unchecked Sendable {
+  @available(*, deprecated, message: "Use instance-based approach with OpenTelemetryConfiguration")
+  public static let instance = DefaultTracer()
+
+  public init() {}
 
   public func spanBuilder(spanName: String) -> SpanBuilder {
     return PropagatedSpanBuilder(tracer: self, spanName: spanName)
